Testing the drag feature. Grid map generator completed
Impleted the first pass of the drag script. You are able to and hold on the cube in the scene and drag it around. It is a bit weird right where it placed the cube higher than the cursor, I'll try and work that out. The grid is also complete and able to generate as many N x M tiles as we want and can specify a distance between them.