better drag feature, can drop an object into place.
Fixed the drag function for the cube to stick with the cursor position instead of being higher than the cursor. The cube now gets a shade of green when it is able to be placed on the grid. Finally, the cube can also now take the place of a grid piece.