grab a cube and place it on the grid and it stays there.
You can now grab a copy of the cube and place it on the grid. Can no longer grab a placed cube, and the grid responds properly to an object being placed there so the shade of cube will be red when there is already a cube there. The cube grabbed from will stay there and you can keep "drawing" from the cube.