Can now tell if there is cube in a position
State manager keeps track of an array that store integers. If the int at that position is a 0 then there is no block there on the grid. Else it will be a 1 and there will be a block there. Next thing needed for implementation will be knowing what type of block was placed there.