can spawn blocks from UI
drag and drop still allows the block to be re placed, but properly calculates the score. there are just blocks of UI that spawn the blocks according to colors I picked, they are also named in the scene. The blocks don't spawn on the mouse location yet, just somewhere in the world.