tiles no longer spawn behind UI
Tiles now spawn in one part of the screen. The player cannot spawn more than one tile at a time, and if the tile has not been picked for some amount of time, the tile will disappear and they can then pick a new one. This still works on rotation.