New Lighting and Settings menu
The settings menu components are better aligned. April and I changed the angle of lighting so that the default point of view doesn't have the blocks in shadow, we also changed the ambient light to be a bit lighter in color, so that when the camera is rotated to the back, its not as dark from the shadows.